wrap markdown module for root config in dsmarkdown module
