+++ src/app/app.module.ts
@@ -2,7 +2,6 @@ import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { FormsModule } from '@angular/forms';
 import { MarkdownModule, MarkedOptions } from 'ngx-markdown';
 
 import { LuModule } from '@lucca-front/ng';
@@ -17,8 +16,7 @@ import { DsNavSideModule } from './nav-side';
 import { DsHomeModule } from './home';
 
 import { AppComponent } from './app.component';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
-import { markedOptionsFactory } from './markdown/markedOptions';
+import { DsMarkdownModule } from './commons';
 
 
 
@@ -27,7 +25,6 @@ import { markedOptionsFactory } from './markdown/markedOptions';
 		BrowserModule,
 		BrowserAnimationsModule,
 		LuModule,
-		HttpClientModule,
 
 		DsHomeModule,
 		DsMainModule,
@@ -37,13 +34,7 @@ import { markedOptionsFactory } from './markdown/markedOptions';
 		DsVisualModule,
 		DsNavSideModule,
 
-		MarkdownModule.forRoot({
-			loader: HttpClient,
-			markedOptions: {
-				provide: MarkedOptions,
-				useFactory: markedOptionsFactory
-			}
-		}),
+		DsMarkdownModule.forRoot(),
 
 		RouterModule.forRoot([], {
 			scrollPositionRestoration: 'enabled',
