+++ src/app/commons/markdown/markdown.module.ts
@@ -1,7 +1,32 @@
-import { NgModule } from '@angular/core';
-import { MarkdownModule } from 'ngx-markdown';
+import { NgModule, ModuleWithProviders } from '@angular/core';
 import { DsMarkdownSrcDirective } from './markdown-src.directive';
 import { DsAssetModule } from '../asset';
+import { MarkdownModule, MarkedOptions } from 'ngx-markdown';
+import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { markedOptionsFactory } from './markdown.options';
+
+@NgModule({
+	imports: [
+		HttpClientModule,
+		MarkdownModule.forRoot({
+			loader: HttpClient,
+			markedOptions: {
+				provide: MarkedOptions,
+				useFactory: markedOptionsFactory
+			}
+		}),
+	],
+	// exports: [
+	// 	MarkdownModule.forRoot({
+	// 		// loader: HttpClient,
+	// 		// markedOptions: {
+	// 		// 	provide: MarkedOptions,
+	// 		// 	useFactory: markedOptionsFactory,
+	// 		// }
+	// 	}),
+	// ],
+})
+class DsMarkdownRootModule {}
 
 @NgModule({
 	imports: [
@@ -16,4 +41,10 @@ import { DsAssetModule } from '../asset';
 		MarkdownModule,
 	],
 })
-export class DsMarkdownModule {}
+export class DsMarkdownModule {
+	public static forRoot(): ModuleWithProviders {
+		return {
+			ngModule: DsMarkdownRootModule,
+		};
+	}
+}
